More code tweaking

+++ Assets/Scripts/ScoreUpdate.cs
@@ -5,14 +5,14 @@ using TMPro;
 
 public class ScoreUpdate : MonoBehaviour
 {
-    [SerializeField] public int scoreTotal = 0;
-    [SerializeField] public int scoreBuffer = 0;
-    [SerializeField] private int previousScoreGain = 0;
+    [SerializeField] public float scoreTotal = 0;
+    [SerializeField] public float scoreBuffer = 0;
+    [SerializeField] private float previousScoreGain = 0;
     [SerializeField] private TextMeshProUGUI scoreGainUI;
     [SerializeField] private TextMeshProUGUI scoreText;
     [SerializeField] private Animator anim;
     [SerializeField] public bool doublePoints;
-    [SerializeField] public int pointsMultiplier;
+    [SerializeField] public float pointsMultiplier;
 
     private void Start()
     {
@@ -51,7 +51,7 @@ public class ScoreUpdate : MonoBehaviour
         Invoke("ResetBuffer", .1f);
     }
 
-    public void CalculateScore(int scoreWorth)
+    public void CalculateScore(float scoreWorth)
     {
         scoreBuffer += scoreWorth;
         if (IsInvoking("UpdateScoreGain"))
